Move BackendConfigDetails from ldap dir to directoryService dir
