+++ graylog2-web-interface/src/components/authentication/directoryServices/BackendConfigDetails/UserSyncSection.jsx
@@ -9,7 +9,7 @@ import SectionComponent from 'components/common/Section/SectionComponent';
 
 import EditLinkButton from './EditLinkButton';
 
-import { STEP_KEY as USER_SYNC_KEY } from '../../BackendWizard/UserSyncStep';
+import { STEP_KEY as USER_SYNC_KEY } from '../BackendWizard/UserSyncStep';
 
 const RolesList = ({ defaultRolesIds, roles }: {defaultRolesIds: Immutable.List<string>, roles: Immutable.List<Role>}) => {
   const defaultRolesNames = defaultRolesIds.map((roleId) => roles.find((role) => role.id === roleId)?.name ?? 'Role not found');
